split substate code from state
