+++ crates/bevy_ecs/src/lib.rs
@@ -40,9 +40,9 @@ pub mod prelude {
         schedule::{
             apply_state_transition, apply_system_buffers, common_conditions::*, Condition,
             IntoSystemConfigs, IntoSystemSet, IntoSystemSetConfig, IntoSystemSetConfigs, NextState,
-            OnEnter, OnExit, OnTransition, OnUpdate, Schedule, Schedules, State, States,
-            SubstateInFn, SubstateLabelInFn, SystemSet,
+            OnEnter, OnExit, OnTransition, OnUpdate, Schedule, Schedules, State, States, SystemSet,
         },
+        
         system::{
             adapter as system_adapter,
             adapter::{dbg, error, ignore, info, unwrap, warn},
