+++ crates/bevy_ecs/src/schedule/state.rs
@@ -1,6 +1,6 @@
 use std::fmt::Debug;
 use std::hash::Hash;
-use std::marker::PhantomData;
+
 use std::mem;
 
 use crate as bevy_ecs;
@@ -10,7 +10,6 @@ use crate::system::Resource;
 use crate::world::World;
 
 pub use bevy_ecs_macros::States;
-use bevy_utils::all_tuples;
 
 /// Types that can define world-wide states in a finite-state machine.
 ///
@@ -56,36 +55,6 @@ pub struct OnEnter<S: States>(pub S);
 #[derive(ScheduleLabel, Clone, Debug, PartialEq, Eq, Hash)]
 pub struct OnExit<S: States>(pub S);
 
-#[derive(ScheduleLabel, Clone, Debug, PartialEq, Eq, Hash)]
-/// Like [AnySubstate], however this one only matches functions that return
-///
-/// This should be able to be used on some unnamed fields in enums. However, please beware that fields that have the same substate might pass the type checking, as they really match type-wise. (
-/// So
-/// ```rs
-/// #[derive(States, Clone, Copy, Default, Eq, PartialEq, Hash, Debug)]
-/// pub enum MyState {
-///     Foo(AliceOrBob)
-///     Bar(AliceOrBob)
-/// }
-///
-/// pub AliceOrBob {
-///     Alice,
-///     Bob
-/// }
-/// // these should compile
-/// let mut foo: Fn(AliceOrBob) -> MyState = MyState::Foo;
-/// foo = MyState::Bar;
-///
-/// ```
-pub struct SubstateLabelInFn<Ret, Args, L: ScheduleLabel>(PhantomData<(L, Ret, Args)>);
-
-pub struct SubstateInFn<L: ScheduleLabel>(PhantomData<L>);
-impl<L: ScheduleLabel> SubstateInFn<L> {
-    /// from the type of F from the value passed into the function
-    pub fn new<F: VariadicFn<Ret, Args>, Ret, Args>(_f: &F) -> SubstateLabelInFn<Ret, Args, L> {
-        SubstateLabelInFn(PhantomData)
-    }
-}
 
 /*
 #[derive(Clone, PartialEq, Eq, Hash, Debug, Default, ScheduleLabel)]
@@ -174,25 +143,3 @@ pub fn apply_state_transition<S: States>(world: &mut World) {
     }
 }
 
-///
-/// A trait which most variadic functions should technically implement?
-/// The issue with Fn trait types is that it doesn't support variadic functions.  This one should
-pub trait VariadicFn<Ret, Args> {}
-
-macro_rules! impl_substate_tuple {
-    ($($name: tt),*)  => {
-        impl<Function, Ret, $( $name, )* >  VariadicFn<Ret, ($($name,)*)> for Function where Function: Fn($($name,)*) -> Ret {}
-    }
-}
-
-pub trait Arguments {}
-macro_rules! impl_args_tuple {
-    ($($name: tt),*) => {
-        impl< $( $name, )* > Arguments for  ($( $name, )* ) {
-
-        }
-    }
-}
-
-all_tuples!(impl_args_tuple, 0, 16, A);
-all_tuples!(impl_substate_tuple, 0, 16, S);
